Issue #25735 move messages to share bundle rather than duplicating them
