--- dev/io.openliberty.concurrent.internal.cdi.4.0/resources/io/openliberty/concurrent/internal/cdi4/resources/CWWKCMessages_de.nlsprops
@@ -1,43 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.concurrent.internal.cdi/resources/io/openliberty/concurrent/internal/cdi/resources/CWWKCMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-#
-#COMPONENTPREFIX CWWKC
-#COMPONENTNAMEFOR CWWKC Jakarta EE Concurrency
-#
-# NLS_MESSAGEFORMAT_VAR
-#
-#   Strings in this file which contain replacement variables are processed by the MessageFormat 
-#   class (single quote must be coded as 2 consecutive single quotes ''). Strings in this file 
-#   which do NOT contain replacement variables are NOT processed by the MessageFormat class 
-#   (single quote must be coded as one single quote '). 
-
-# All messages must use the range CWWCK1400 to CWWCK1499
-
-CWWKC1400.unsupported.return.type=CWWKC1400E: {0} ist kein g\u00fcltiger R\u00fcckgabetyp f\u00fcr die Methode {1} der Klasse {2}. G\u00fcltige R\u00fcckgabetypen f\u00fcr Methoden, die mit {3} annotiert sind: {4}.
-CWWKC1400.unsupported.return.type.explanation=Die Spezifikation Jakarta Concurrency begrenzt den Bereich der g\u00fcltigen R\u00fcckgabetypen f\u00fcr Methoden, die diese Annotation verwenden.
-CWWKC1400.unsupported.return.type.useraction=Aktualisieren Sie die Methode in der Anwendung, um zu einem der unterst\u00fctzten R\u00fcckgabetypen zu wechseln, oder entfernen Sie die Annotation aus der Methode.
-
-CWWKC1401.class.anno.disallowed=CWWKC1401E: Die Annotation {0} ist auf Methodenebene zul\u00e4ssig. Sie kann nicht auf Klassenebene in der Klasse {1} verwendet werden.
-CWWKC1401.class.anno.disallowed.explanation=Diese Annotation darf nicht auf Klassenebene angegeben werden.
-CWWKC1401.class.anno.disallowed.useraction=Verschieben Sie die Annotation in die einzelnen Methoden der Klasse.
-
-CWWKC1402.not.managed.executor=CWWKC1402E: Die Annotation " {0} " in der Methode {1} der Klasse {2} gibt den JNDI-Namen {3} an, der in die Ressource {4} aufgel\u00f6st wird, die kein ManagedExecutorService oder ManagedScheduledExecutorService ist. Von der Ressource implementierte Schnittstellen: {5}.
-CWWKC1402.not.managed.executor.explanation=Der JNDI-Name muss in eine Ressource aufgel\u00f6st werden, die die Schnittstelle "ManagedExecutorService" oder "ManagedScheduledExecutorService" implementiert.
-CWWKC1402.not.managed.executor.useraction=Korrigieren Sie den JNDI-Namen, um eine Ressource anzugeben, auf die die Anwendungskomponente zugreifen kann und die die Schnittstelle "ManagedExecutorService" oder die Schnittstelle "ManagedScheduledExecutorService" implementiert.
-
-CWWKC1403.unsupported.tx.type=CWWKC1403E: Die Annotation {0} gibt den Wert {1} an, der in Kombination mit der Annotation {2}, die sich in der Methode {3} der Klasse {4} befindet, nicht zul\u00e4ssig ist. Zul\u00e4ssige Werte: {5}.
-CWWKC1403.unsupported.tx.type.explanation=Die gemeinsame Verwendung von zwei Annotationen h\u00e4ngt davon ab, ob der Wert der ersten angegebenen Annotation mit der zweiten Annotation zul\u00e4ssig ist.
-CWWKC1403.unsupported.tx.type.useraction=Aktualisieren Sie die Anwendung, um zu einem der unterst\u00fctzten Werte zu wechseln, oder entfernen Sie eine der Anmerkungen.
