--- dev/io.openliberty.concurrent.internal.cdi.4.0/resources/io/openliberty/concurrent/internal/cdi4/resources/CWWKCMessages_es.nlsprops
@@ -1,43 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.concurrent.internal.cdi/resources/io/openliberty/concurrent/internal/cdi/resources/CWWKCMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-#
-#COMPONENTPREFIX CWWKC
-#COMPONENTNAMEFOR CWWKC Jakarta EE Concurrency
-#
-# NLS_MESSAGEFORMAT_VAR
-#
-#   Strings in this file which contain replacement variables are processed by the MessageFormat 
-#   class (single quote must be coded as 2 consecutive single quotes ''). Strings in this file 
-#   which do NOT contain replacement variables are NOT processed by the MessageFormat class 
-#   (single quote must be coded as one single quote '). 
-
-# All messages must use the range CWWCK1400 to CWWCK1499
-
-CWWKC1400.unsupported.return.type=CWWKC1400E: {0} no es un tipo de retorno v\u00e1lido para el m\u00e9todo {1} de la clase {2}. Los tipos de retorno v\u00e1lidos para los m\u00e9todos anotados con {3} son: {4}.
-CWWKC1400.unsupported.return.type.explanation=La especificaci\u00f3n de simultaneidad de Jakarta limita el rango de tipos de retorno v\u00e1lidos para los m\u00e9todos que tienen esta anotaci\u00f3n.
-CWWKC1400.unsupported.return.type.useraction=Actualice el m\u00e9todo dentro de la aplicaci\u00f3n para conmutar a uno de los tipos de retorno admitidos o elimine la anotaci\u00f3n del m\u00e9todo.
-
-CWWKC1401.class.anno.disallowed=CWWKC1401E: La anotaci\u00f3n {0} est\u00e1 permitida a nivel de m\u00e9todo. No se puede utilizar a nivel de clase en la clase {1}.
-CWWKC1401.class.anno.disallowed.explanation=Esta anotaci\u00f3n no se debe especificar a nivel de clase.
-CWWKC1401.class.anno.disallowed.useraction=Mueva la anotaci\u00f3n a los m\u00e9todos individuales de la clase.
-
-CWWKC1402.not.managed.executor=CWWKC1402E: La anotaci\u00f3n {0} en el m\u00e9todo {1} de la clase {2} especifica el nombre de JNDI {3}, que se resuelve en el recurso {4} que no es un ManagedExecutorService o ManagedScheduledExecutorService. Las interfaces que implementa el recurso son: {5}.
-CWWKC1402.not.managed.executor.explanation=El nombre de JNDI se debe resolver en un recurso que implemente la interfaz ManagedExecutorService o la interfaz ManagedScheduledExecutorService.
-CWWKC1402.not.managed.executor.useraction=Corrija el nombre de JNDI para que especifique un recurso que sea accesible para el componente de aplicaci\u00f3n y que implemente la interfaz ManagedExecutorService o la interfaz ManagedScheduledExecutorService.
-
-CWWKC1403.unsupported.tx.type=CWWKC1403E: La anotaci\u00f3n {0} especifica el valor {1}, que no est\u00e1 permitido en combinaci\u00f3n con la anotaci\u00f3n {2} que est\u00e1 en el m\u00e9todo {3} de la clase {4}. Los valores permitidos son: {5}.
-CWWKC1403.unsupported.tx.type.explanation=El uso de dos anotaciones de forma conjunta depende de si el valor de la primera anotaci\u00f3n que se especifica est\u00e1 permitido con la segunda anotaci\u00f3n.
-CWWKC1403.unsupported.tx.type.useraction=Actualice la aplicaci\u00f3n para conmutar a uno de los valores admitidos o elimine una de las anotaciones.
