--- dev/io.openliberty.concurrent.internal.cdi.4.0/resources/io/openliberty/concurrent/internal/cdi4/resources/CWWKCMessages_ja.nlsprops
@@ -1,43 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.concurrent.internal.cdi/resources/io/openliberty/concurrent/internal/cdi/resources/CWWKCMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-#
-#COMPONENTPREFIX CWWKC
-#COMPONENTNAMEFOR CWWKC Jakarta EE Concurrency
-#
-# NLS_MESSAGEFORMAT_VAR
-#
-#   Strings in this file which contain replacement variables are processed by the MessageFormat 
-#   class (single quote must be coded as 2 consecutive single quotes ''). Strings in this file 
-#   which do NOT contain replacement variables are NOT processed by the MessageFormat class 
-#   (single quote must be coded as one single quote '). 
-
-# All messages must use the range CWWCK1400 to CWWCK1499
-
-CWWKC1400.unsupported.return.type=CWWKC1400E: {0} \u306f\u3001 {2} \u30af\u30e9\u30b9\u306e {1} \u30e1\u30bd\u30c3\u30c9\u306e\u6709\u52b9\u306a\u623b\u308a\u306e\u578b\u3067\u306f\u3042\u308a\u307e\u305b\u3093\u3002 {3} \u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u304c\u4ed8\u3051\u3089\u308c\u305f\u30e1\u30bd\u30c3\u30c9\u306e\u6709\u52b9\u306a\u623b\u308a\u306e\u578b\u306f\u3001 {4}\u3067\u3059\u3002
-CWWKC1400.unsupported.return.type.explanation=Jakarta Concurrency \u4ed5\u69d8\u306f\u3001\u3053\u306e\u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u3092\u6301\u3064\u30e1\u30bd\u30c3\u30c9\u306e\u6709\u52b9\u306a\u623b\u308a\u306e\u578b\u306e\u7bc4\u56f2\u3092\u5236\u9650\u3057\u307e\u3059\u3002
-CWWKC1400.unsupported.return.type.useraction=\u30a2\u30d7\u30ea\u30b1\u30fc\u30b7\u30e7\u30f3\u5185\u306e\u30e1\u30bd\u30c3\u30c9\u3092\u66f4\u65b0\u3057\u3066\u3001\u30b5\u30dd\u30fc\u30c8\u3055\u308c\u3066\u3044\u308b\u3044\u305a\u308c\u304b\u306e\u623b\u308a\u306e\u578b\u306b\u5207\u308a\u66ff\u3048\u308b\u304b\u3001\u30e1\u30bd\u30c3\u30c9\u304b\u3089\u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u3092\u524a\u9664\u3057\u3066\u304f\u3060\u3055\u3044\u3002
-
-CWWKC1401.class.anno.disallowed=CWWKC1401E: {0} \u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u306f\u3001\u30e1\u30bd\u30c3\u30c9\u30fb\u30ec\u30d9\u30eb\u3067\u8a31\u53ef\u3055\u308c\u307e\u3059\u3002 {1} \u30af\u30e9\u30b9\u306e\u30af\u30e9\u30b9\u30fb\u30ec\u30d9\u30eb\u3067\u306f\u4f7f\u7528\u3067\u304d\u307e\u305b\u3093\u3002
-CWWKC1401.class.anno.disallowed.explanation=\u3053\u306e\u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u306f\u3001\u30af\u30e9\u30b9\u30fb\u30ec\u30d9\u30eb\u3067\u6307\u5b9a\u3057\u3066\u306f\u306a\u308a\u307e\u305b\u3093\u3002
-CWWKC1401.class.anno.disallowed.useraction=\u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u3092\u30af\u30e9\u30b9\u306e\u500b\u3005\u306e\u30e1\u30bd\u30c3\u30c9\u306b\u79fb\u52d5\u3057\u307e\u3059\u3002
-
-CWWKC1402.not.managed.executor=CWWKC1402E: {2} \u30af\u30e9\u30b9\u306e {1} \u30e1\u30bd\u30c3\u30c9\u306e {0} \u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u306f\u3001 {3} JNDI \u540d\u3092\u6307\u5b9a\u3057\u307e\u3059\u3002\u3053\u308c\u306f\u3001ManagedExecutorService \u3067\u3082 ManagedScheduledExecutorService \u3067\u3082\u306a\u3044 {4} \u30ea\u30bd\u30fc\u30b9\u306b\u89e3\u6c7a\u3055\u308c\u307e\u3059\u3002 \u30ea\u30bd\u30fc\u30b9\u306b\u3088\u3063\u3066\u5b9f\u88c5\u3055\u308c\u308b\u30a4\u30f3\u30bf\u30fc\u30d5\u30a7\u30fc\u30b9\u306f\u3001 {5}\u3067\u3059\u3002
-CWWKC1402.not.managed.executor.explanation=JNDI \u540d\u306f\u3001ManagedExecutorService \u30a4\u30f3\u30bf\u30fc\u30d5\u30a7\u30fc\u30b9\u307e\u305f\u306f ManagedScheduledExecutorService \u30a4\u30f3\u30bf\u30fc\u30d5\u30a7\u30fc\u30b9\u3092\u5b9f\u88c5\u3059\u308b\u30ea\u30bd\u30fc\u30b9\u306b\u89e3\u6c7a\u3059\u308b\u5fc5\u8981\u304c\u3042\u308a\u307e\u3059\u3002
-CWWKC1402.not.managed.executor.useraction=JNDI \u540d\u3092\u8a02\u6b63\u3057\u3066\u3001\u30a2\u30d7\u30ea\u30b1\u30fc\u30b7\u30e7\u30f3\u30fb\u30b3\u30f3\u30dd\u30fc\u30cd\u30f3\u30c8\u306b\u30a2\u30af\u30bb\u30b9\u53ef\u80fd\u3067\u3001ManagedExecutorService \u30a4\u30f3\u30bf\u30fc\u30d5\u30a7\u30fc\u30b9\u307e\u305f\u306f ManagedScheduledExecutorService \u30a4\u30f3\u30bf\u30fc\u30d5\u30a7\u30fc\u30b9\u3092\u5b9f\u88c5\u3059\u308b\u30ea\u30bd\u30fc\u30b9\u3092\u6307\u5b9a\u3057\u3066\u304f\u3060\u3055\u3044\u3002
-
-CWWKC1403.unsupported.tx.type=CWWKC1403E: {0} \u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u306b {1} \u5024\u304c\u6307\u5b9a\u3055\u308c\u3066\u3044\u307e\u3059\u304c\u3001 {4} \u30af\u30e9\u30b9\u306e {3} \u30e1\u30bd\u30c3\u30c9\u306b\u3042\u308b {2} \u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u3068\u7d44\u307f\u5408\u308f\u305b\u3066\u4f7f\u7528\u3059\u308b\u3053\u3068\u306f\u8a31\u53ef\u3055\u308c\u3066\u3044\u307e\u305b\u3093\u3002 \u8a31\u53ef\u3055\u308c\u308b\u5024\u306f: {5}\u3067\u3059\u3002
-CWWKC1403.unsupported.tx.type.explanation=2 \u3064\u306e\u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u3092\u4e00\u7dd2\u306b\u4f7f\u7528\u3059\u308b\u304b\u3069\u3046\u304b\u306f\u3001\u6307\u5b9a\u3055\u308c\u305f\u6700\u521d\u306e\u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u306e\u5024\u304c 2 \u756a\u76ee\u306e\u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u3067\u8a31\u53ef\u3055\u308c\u3066\u3044\u308b\u304b\u3069\u3046\u304b\u306b\u3088\u3063\u3066\u6c7a\u307e\u308a\u307e\u3059\u3002
-CWWKC1403.unsupported.tx.type.useraction=\u30a2\u30d7\u30ea\u30b1\u30fc\u30b7\u30e7\u30f3\u3092\u66f4\u65b0\u3057\u3066\u3001\u30b5\u30dd\u30fc\u30c8\u3055\u308c\u3066\u3044\u308b\u5024\u306e 1 \u3064\u306b\u5207\u308a\u66ff\u3048\u308b\u304b\u3001\u30a2\u30ce\u30c6\u30fc\u30b7\u30e7\u30f3\u306e 1 \u3064\u3092\u524a\u9664\u3057\u3066\u304f\u3060\u3055\u3044\u3002
