--- dev/io.openliberty.concurrent.internal.cdi.4.0/resources/io/openliberty/concurrent/internal/cdi4/resources/CWWKCMessages_ko.nlsprops
@@ -1,43 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.concurrent.internal.cdi/resources/io/openliberty/concurrent/internal/cdi/resources/CWWKCMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-#
-#COMPONENTPREFIX CWWKC
-#COMPONENTNAMEFOR CWWKC Jakarta EE Concurrency
-#
-# NLS_MESSAGEFORMAT_VAR
-#
-#   Strings in this file which contain replacement variables are processed by the MessageFormat 
-#   class (single quote must be coded as 2 consecutive single quotes ''). Strings in this file 
-#   which do NOT contain replacement variables are NOT processed by the MessageFormat class 
-#   (single quote must be coded as one single quote '). 
-
-# All messages must use the range CWWCK1400 to CWWCK1499
-
-CWWKC1400.unsupported.return.type=CWWKC1400E: {0}\uc740(\ub294) {2} \ud074\ub798\uc2a4\uc758 {1} \uba54\uc18c\ub4dc\uc5d0 \ub300\ud574 \uc720\ud6a8\ud55c \ub9ac\ud134 \uc720\ud615\uc774 \uc544\ub2d9\ub2c8\ub2e4. {3}\uc758 \uc5b4\ub178\ud14c\uc774\uc158\uc774 \uc788\ub294 \uba54\uc18c\ub4dc\uc5d0 \ub300\ud574 \uc720\ud6a8\ud55c \ub9ac\ud134 \uc720\ud615\uc740 {4}\uc785\ub2c8\ub2e4.
-CWWKC1400.unsupported.return.type.explanation=Jakarta \ub3d9\uc2dc\uc131 \uc2a4\ud399\uc740 \uc774 \uc5b4\ub178\ud14c\uc774\uc158\uc774 \uc788\ub294 \uba54\uc18c\ub4dc\uc5d0 \ub300\ud574 \uc720\ud6a8\ud55c \ub9ac\ud134 \uc720\ud615\uc758 \ubc94\uc704\ub97c \uc81c\ud55c\ud569\ub2c8\ub2e4.
-CWWKC1400.unsupported.return.type.useraction=\uc560\ud50c\ub9ac\ucf00\uc774\uc158 \ub0b4\uc758 \uba54\uc18c\ub4dc\ub97c \uc5c5\ub370\uc774\ud2b8\ud558\uc5ec \uc9c0\uc6d0\ub418\ub294 \ub9ac\ud134 \uc720\ud615 \uc911 \ud558\ub098\ub85c \uc804\ud658\ud558\uac70\ub098 \uba54\uc18c\ub4dc\uc5d0\uc11c \uc5b4\ub178\ud14c\uc774\uc158\uc744 \uc81c\uac70\ud558\uc2ed\uc2dc\uc624.
-
-CWWKC1401.class.anno.disallowed=CWWKC1401E: {0} \uc5b4\ub178\ud14c\uc774\uc158\uc740 \uba54\uc18c\ub4dc \ub808\ubca8\uc5d0\uc11c \ud5c8\uc6a9\ub429\ub2c8\ub2e4. {1} \ud074\ub798\uc2a4\uc758 \ud074\ub798\uc2a4 \ub808\ubca8\uc5d0\uc11c\ub294 \uc0ac\uc6a9\ud560 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4.
-CWWKC1401.class.anno.disallowed.explanation=\uc774 \uc5b4\ub178\ud14c\uc774\uc158\uc740 \ud074\ub798\uc2a4 \ub808\ubca8\uc5d0\uc11c \uc9c0\uc815\ud560 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4.
-CWWKC1401.class.anno.disallowed.useraction=\uc5b4\ub178\ud14c\uc774\uc158\uc744 \ud074\ub798\uc2a4\uc758 \uac1c\ubcc4 \uba54\uc18c\ub4dc\ub85c \uc774\ub3d9\ud558\uc2ed\uc2dc\uc624.
-
-CWWKC1402.not.managed.executor=CWWKC1402E: {2} \ud074\ub798\uc2a4\uc758 {1} \uba54\uc18c\ub4dc\uc5d0 \ub300\ud55c {0} \uc5b4\ub178\ud14c\uc774\uc158\uc740 {3} JNDI \uc774\ub984\uc744 \uc9c0\uc815\ud558\uba70, \uc774\ub294 ManagedExecutorService \ub610\ub294 ManagedScheduledExecutorService\uac00 \uc544\ub2cc {4} \uc790\uc6d0\uc73c\ub85c \ubd84\uc11d\ub429\ub2c8\ub2e4. \uc790\uc6d0\uc5d0\uc11c \uad6c\ud604\ub41c \uc778\ud130\ud398\uc774\uc2a4\ub294 {5}\uc785\ub2c8\ub2e4.
-CWWKC1402.not.managed.executor.explanation=JNDI \uc774\ub984\uc740 ManagedExecutorService \uc778\ud130\ud398\uc774\uc2a4 \ub610\ub294 ManagedScheduledExecutorService \uc778\ud130\ud398\uc774\uc2a4\ub97c \uad6c\ud604\ud558\ub294 \uc790\uc6d0\uc73c\ub85c \ubd84\uc11d\ub418\uc5b4\uc57c \ud569\ub2c8\ub2e4.
-CWWKC1402.not.managed.executor.useraction=JNDI \uc774\ub984\uc744 \uc815\uc815\ud558\uc5ec \uc560\ud50c\ub9ac\ucf00\uc774\uc158 \ucef4\ud3ec\ub10c\ud2b8\uc5d0 \uc561\uc138\uc2a4\ud560 \uc218 \uc788\uace0 ManagedExecutorService \uc778\ud130\ud398\uc774\uc2a4 \ub610\ub294 ManagedScheduledExecutorService \uc778\ud130\ud398\uc774\uc2a4\ub97c \uad6c\ud604\ud558\ub294 \uc790\uc6d0\uc744 \uc9c0\uc815\ud558\uc2ed\uc2dc\uc624.
-
-CWWKC1403.unsupported.tx.type=CWWKC1403E: {0} \uc5b4\ub178\ud14c\uc774\uc158\uc774 {4} \ud074\ub798\uc2a4\uc758 {3} \uba54\uc18c\ub4dc\uc5d0 \uc788\ub294 {2} \uc5b4\ub178\ud14c\uc774\uc158\uacfc \ud568\uaed8 \uc0ac\uc6a9\ud560 \uc218 \uc5c6\ub294 {1} \uac12\uc744 \uc9c0\uc815\ud569\ub2c8\ub2e4. \ud5c8\uc6a9\ub41c \uac12\uc740 {5}\uc785\ub2c8\ub2e4.
-CWWKC1403.unsupported.tx.type.explanation=\ub450 \uac1c\uc758 \uc5b4\ub178\ud14c\uc774\uc158\uc744 \ud568\uaed8 \uc0ac\uc6a9\ud558\ub294 \uac83\uc740 \uc9c0\uc815\ub41c \uccab \ubc88\uc9f8 \uc5b4\ub178\ud14c\uc774\uc158\uc758 \uac12\uc774 \ub450 \ubc88\uc9f8 \uc5b4\ub178\ud14c\uc774\uc158\uc5d0\uc11c \ud5c8\uc6a9\ub418\ub294\uc9c0 \uc5ec\ubd80\uc5d0 \ub530\ub77c \ub2e4\ub985\ub2c8\ub2e4.
-CWWKC1403.unsupported.tx.type.useraction=\uc9c0\uc6d0\ub418\ub294 \uac12 \uc911 \ud558\ub098\ub85c \uc804\ud658\ud558\uac70\ub098 \uc5b4\ub178\ud14c\uc774\uc158 \uc911 \ud558\ub098\ub97c \uc81c\uac70\ud558\ub3c4\ub85d \uc560\ud50c\ub9ac\ucf00\uc774\uc158\uc744 \uc5c5\ub370\uc774\ud2b8\ud558\uc2ed\uc2dc\uc624.
