--- dev/io.openliberty.concurrent.internal.cdi.4.0/resources/io/openliberty/concurrent/internal/cdi4/resources/CWWKCMessages_pt_BR.nlsprops
@@ -1,43 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.concurrent.internal.cdi/resources/io/openliberty/concurrent/internal/cdi/resources/CWWKCMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-#
-#COMPONENTPREFIX CWWKC
-#COMPONENTNAMEFOR CWWKC Jakarta EE Concurrency
-#
-# NLS_MESSAGEFORMAT_VAR
-#
-#   Strings in this file which contain replacement variables are processed by the MessageFormat 
-#   class (single quote must be coded as 2 consecutive single quotes ''). Strings in this file 
-#   which do NOT contain replacement variables are NOT processed by the MessageFormat class 
-#   (single quote must be coded as one single quote '). 
-
-# All messages must use the range CWWCK1400 to CWWCK1499
-
-CWWKC1400.unsupported.return.type=CWWKC1400E: {0} n\u00e3o \u00e9 um tipo de retorno v\u00e1lido para o m\u00e9todo {1} da classe {2}. Tipos de retorno v\u00e1lidos para m\u00e9todos que s\u00e3o anotados com {3} s\u00e3o: {4}.
-CWWKC1400.unsupported.return.type.explanation=A especifica\u00e7\u00e3o de simultaneidade de Jakarta limita o intervalo de tipos de retorno v\u00e1lidos para m\u00e9todos que possuem essa anota\u00e7\u00e3o.
-CWWKC1400.unsupported.return.type.useraction=Atualize o m\u00e9todo no aplicativo para alternar para um dos tipos de retorno com suporte ou remova a anota\u00e7\u00e3o do m\u00e9todo.
-
-CWWKC1401.class.anno.disallowed=CWWKC1401E: A anota\u00e7\u00e3o {0} \u00e9 permitida no n\u00edvel do m\u00e9todo. Ela n\u00e3o pode ser usada no n\u00edvel de classe na classe {1}.
-CWWKC1401.class.anno.disallowed.explanation=Esta anota\u00e7\u00e3o n\u00e3o deve ser especificada no n\u00edvel de classe.
-CWWKC1401.class.anno.disallowed.useraction=Mova a anota\u00e7\u00e3o para os m\u00e9todos individuais da classe.
-
-CWWKC1402.not.managed.executor=CWWKC1402E: A anota\u00e7\u00e3o {0} no m\u00e9todo {1} da classe {2} especifica o nome de JNDI {3}, que resolve para o recurso {4} que n\u00e3o \u00e9 um ManagedExecutorService ou ManagedProgramedExecutorService. As interfaces implementadas pelo recurso s\u00e3o: {5}.
-CWWKC1402.not.managed.executor.explanation=O nome de JNDI deve resolver para um recurso que implementa a interface ManagedExecutorService ou a interface ManagedProgramedExecutorService.
-CWWKC1402.not.managed.executor.useraction=Corrija o nome de JNDI para especificar um recurso acess\u00edvel ao componente do aplicativo e implemento da interface ManagedExecutorService ou da interface ManagedExecutorService.
-
-CWWKC1403.unsupported.tx.type=CWWKC1403E: A anota\u00e7\u00e3o {0} especifica o valor {1}, o que n\u00e3o \u00e9 permitido em combina\u00e7\u00e3o com a anota\u00e7\u00e3o {2} que est\u00e1 no m\u00e9todo {3} da classe {4}. Os valores permitidos s\u00e3o: {5}.
-CWWKC1403.unsupported.tx.type.explanation=O uso de duas anota\u00e7\u00f5es juntas depende se o valor da primeira anota\u00e7\u00e3o especificada \u00e9 permitido com a segunda anota\u00e7\u00e3o.
-CWWKC1403.unsupported.tx.type.useraction=Atualize o aplicativo para alternar para um dos valores suportados ou remova uma das anota\u00e7\u00f5es.
