--- dev/io.openliberty.concurrent.internal.cdi.4.0/resources/io/openliberty/concurrent/internal/cdi4/resources/CWWKCMessages_zh.nlsprops
@@ -1,43 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.concurrent.internal.cdi/resources/io/openliberty/concurrent/internal/cdi/resources/CWWKCMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-#
-#COMPONENTPREFIX CWWKC
-#COMPONENTNAMEFOR CWWKC Jakarta EE Concurrency
-#
-# NLS_MESSAGEFORMAT_VAR
-#
-#   Strings in this file which contain replacement variables are processed by the MessageFormat 
-#   class (single quote must be coded as 2 consecutive single quotes ''). Strings in this file 
-#   which do NOT contain replacement variables are NOT processed by the MessageFormat class 
-#   (single quote must be coded as one single quote '). 
-
-# All messages must use the range CWWCK1400 to CWWCK1499
-
-CWWKC1400.unsupported.return.type=CWWKC1400E: \u5bf9\u4e8e {2} \u7c7b\u7684 {1} \u65b9\u6cd5\u800c\u8a00\uff0c{0} \u4e0d\u662f\u6709\u6548\u7684\u8fd4\u56de\u7c7b\u578b\u3002 \u4f7f\u7528 {3} \u6ce8\u91ca\u7684\u65b9\u6cd5\u7684\u6709\u6548\u8fd4\u56de\u7c7b\u578b\u4e3a\uff1a{4}\u3002
-CWWKC1400.unsupported.return.type.explanation=Jakarta \u5e76\u884c\u89c4\u8303\u9650\u5236\u5177\u6709\u6b64\u6ce8\u91ca\u7684\u65b9\u6cd5\u7684\u6709\u6548\u8fd4\u56de\u7c7b\u578b\u8303\u56f4\u3002
-CWWKC1400.unsupported.return.type.useraction=\u66f4\u65b0\u5e94\u7528\u7a0b\u5e8f\u4e2d\u7684\u65b9\u6cd5\u4ee5\u5207\u6362\u5230\u5176\u4e2d\u4e00\u79cd\u53d7\u652f\u6301\u7684\u8fd4\u56de\u7c7b\u578b\uff0c\u6216\u8005\u4ece\u65b9\u6cd5\u4e2d\u79fb\u9664\u6ce8\u91ca\u3002
-
-CWWKC1401.class.anno.disallowed=CWWKC1401E: {0} \u6ce8\u91ca\u5728\u65b9\u6cd5\u7ea7\u522b\u662f\u5141\u8bb8\u7684\u3002 \u4e0d\u80fd\u5728 {1} \u7c7b\u4e0a\u7684\u7c7b\u7ea7\u522b\u4f7f\u7528\u6b64\u53c2\u6570\u3002
-CWWKC1401.class.anno.disallowed.explanation=\u4e0d\u80fd\u5728\u7c7b\u7ea7\u522b\u6307\u5b9a\u8be5\u6ce8\u91ca\u3002
-CWWKC1401.class.anno.disallowed.useraction=\u5c06\u6ce8\u91ca\u79fb\u52a8\u5230\u7c7b\u7684\u4e2a\u522b\u65b9\u6cd5\u3002
-
-CWWKC1402.not.managed.executor=CWWKC1402E: {2} \u7c7b\u7684 {1} \u65b9\u6cd5\u4e0a\u7684 {0} \u6ce8\u91ca\u6307\u5b9a\u4e86 {3} JNDI \u540d\u79f0\uff0c\u8be5\u540d\u79f0\u5c06\u89e3\u6790\u4e3a {4} \u8d44\u6e90\uff0c\u8be5\u8d44\u6e90\u4e0d\u662f ManagedExecutorService \u6216 ManagedScheduledExecutorService\u3002 \u8d44\u6e90\u5b9e\u73b0\u7684\u63a5\u53e3\u4e3a\uff1a{5}\u3002
-CWWKC1402.not.managed.executor.explanation=\u8be5 JNDI \u540d\u79f0\u5fc5\u987b\u89e3\u6790\u4e3a\u5b9e\u73b0 ManagedExecutorService \u63a5\u53e3\u6216 ManagedScheduledExecutorService \u63a5\u53e3\u7684\u8d44\u6e90\u3002
-CWWKC1402.not.managed.executor.useraction=\u8bf7\u66f4\u6b63 JNDI \u540d\u79f0\u4ee5\u6307\u5b9a\u53ef\u4f9b\u5e94\u7528\u7a0b\u5e8f\u7ec4\u4ef6\u8bbf\u95ee\u7684\u8d44\u6e90\uff0c\u5e76\u5b9e\u73b0 ManagedExecutorService \u63a5\u53e3\u6216 ManagedScheduledExecutorService \u63a5\u53e3\u3002
-
-CWWKC1403.unsupported.tx.type=CWWKC1403E: {0} \u6ce8\u91ca\u6307\u5b9a {1} \u503c\uff0c\u4e0d\u5141\u8bb8\u5c06\u8be5\u503c\u4e0e {4} \u7c7b\u7684 {3} \u65b9\u6cd5\u4e0a\u7684 {2} \u6ce8\u91ca\u7ed3\u5408\u4f7f\u7528\u3002 \u5141\u8bb8\u7684\u503c\u4e3a\uff1a{5}\u3002
-CWWKC1403.unsupported.tx.type.explanation=\u5c06\u4e24\u4e2a\u6ce8\u91ca\u4e00\u8d77\u4f7f\u7528\u53d6\u51b3\u4e8e\u6307\u5b9a\u7684\u7b2c\u4e00\u4e2a\u6ce8\u91ca\u7684\u503c\u662f\u5426\u5141\u8bb8\u4e0e\u7b2c\u4e8c\u4e2a\u6ce8\u91ca\u4e00\u8d77\u4f7f\u7528\u3002
-CWWKC1403.unsupported.tx.type.useraction=\u66f4\u65b0\u5e94\u7528\u7a0b\u5e8f\u4ee5\u5207\u6362\u5230\u5176\u4e2d\u4e00\u4e2a\u53d7\u652f\u6301\u7684\u503c\uff0c\u6216\u79fb\u9664\u5176\u4e2d\u4e00\u4e2a\u6ce8\u91ca\u3002
