--- dev/io.openliberty.concurrent.internal.cdi.4.0/resources/io/openliberty/concurrent/internal/cdi4/resources/CWWKCMessages_zh_TW.nlsprops
@@ -1,43 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.concurrent.internal.cdi/resources/io/openliberty/concurrent/internal/cdi/resources/CWWKCMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-#
-#COMPONENTPREFIX CWWKC
-#COMPONENTNAMEFOR CWWKC Jakarta EE Concurrency
-#
-# NLS_MESSAGEFORMAT_VAR
-#
-#   Strings in this file which contain replacement variables are processed by the MessageFormat 
-#   class (single quote must be coded as 2 consecutive single quotes ''). Strings in this file 
-#   which do NOT contain replacement variables are NOT processed by the MessageFormat class 
-#   (single quote must be coded as one single quote '). 
-
-# All messages must use the range CWWCK1400 to CWWCK1499
-
-CWWKC1400.unsupported.return.type=CWWKC1400E: {0} \u4e0d\u662f {2} \u985e\u5225\u4e4b {1} \u65b9\u6cd5\u7684\u6709\u6548\u50b3\u56de\u985e\u578b\u3002 \u4ee5 {3} \u6a19\u8a3b\u4e4b\u65b9\u6cd5\u7684\u6709\u6548\u50b3\u56de\u985e\u578b\u70ba\uff1a{4}\u3002
-CWWKC1400.unsupported.return.type.explanation=Jakarta \u4e26\u884c\u898f\u683c\u6703\u9650\u5236\u5177\u6709\u6b64\u8a3b\u91cb\u4e4b\u65b9\u6cd5\u7684\u6709\u6548\u50b3\u56de\u985e\u578b\u7bc4\u570d\u3002
-CWWKC1400.unsupported.return.type.useraction=\u8acb\u66f4\u65b0\u61c9\u7528\u7a0b\u5f0f\u5167\u7684\u65b9\u6cd5\uff0c\u4ee5\u5207\u63db\u81f3\u5176\u4e2d\u4e00\u500b\u652f\u63f4\u7684\u50b3\u56de\u985e\u578b\uff0c\u6216\u5f9e\u65b9\u6cd5\u4e2d\u79fb\u9664\u6b64\u8a3b\u91cb\u3002
-
-CWWKC1401.class.anno.disallowed=CWWKC1401E: \u5728\u65b9\u6cd5\u5c64\u6b21\u5bb9\u8a31\u4f7f\u7528 {0} \u8a3b\u91cb\u3002 \u5b83\u7121\u6cd5\u5728 {1} \u985e\u5225\u4e0a\u7684\u985e\u5225\u5c64\u6b21\u4f7f\u7528\u3002
-CWWKC1401.class.anno.disallowed.explanation=\u4e0d\u5f97\u5728\u985e\u5225\u5c64\u6b21\u6307\u5b9a\u6b64\u8a3b\u91cb\u3002
-CWWKC1401.class.anno.disallowed.useraction=\u5c07\u8a3b\u91cb\u79fb\u81f3\u985e\u5225\u7684\u500b\u5225\u65b9\u6cd5\u3002
-
-CWWKC1402.not.managed.executor=CWWKC1402E: {2} \u985e\u5225\u7684 {1} \u65b9\u6cd5\u4e0a\u7684 {0} \u8a3b\u91cb\u6307\u5b9a {3} JNDI \u540d\u7a31\uff0c\u5b83\u6703\u89e3\u6790\u70ba\u4e0d\u662f ManagedExecutorService \u6216 ManagedScheduledExecutorService \u7684 {4} \u8cc7\u6e90\u3002 \u8cc7\u6e90\u6240\u5be6\u4f5c\u7684\u4ecb\u9762\u70ba\uff1a{5}\u3002
-CWWKC1402.not.managed.executor.explanation=JNDI \u540d\u7a31\u5fc5\u9808\u89e3\u6790\u70ba\u5be6\u4f5c ManagedExecutorService \u4ecb\u9762\u6216 ManagedScheduledExecutorService \u4ecb\u9762\u7684\u8cc7\u6e90\u3002
-CWWKC1402.not.managed.executor.useraction=\u8acb\u66f4\u6b63 JNDI \u540d\u7a31\uff0c\u4ee5\u6307\u5b9a\u53ef\u4f9b\u61c9\u7528\u7a0b\u5f0f\u5143\u4ef6\u5b58\u53d6\u4e14\u53ef\u5be6\u4f5c ManagedExecutorService \u4ecb\u9762\u6216 ManagedScheduledExecutorService \u4ecb\u9762\u7684\u8cc7\u6e90\u3002
-
-CWWKC1403.unsupported.tx.type=CWWKC1403E: {0} \u8a3b\u91cb\u6307\u5b9a {1} \u503c\uff0c\u8a72\u503c\u4e0d\u5141\u8a31\u8207 {4} \u985e\u5225\u7684 {3} \u65b9\u6cd5\u4e0a\u7684 {2} \u8a3b\u91cb\u4e00\u8d77\u4f7f\u7528\u3002 \u5141\u8a31\u7684\u503c\u70ba\uff1a{5}\u3002
-CWWKC1403.unsupported.tx.type.explanation=\u5169\u500b\u8a3b\u91cb\u80fd\u5426\u4e00\u8d77\u4f7f\u7528\u53d6\u6c7a\u65bc\u7b2c\u4e8c\u500b\u8a3b\u91cb\u662f\u5426\u5141\u8a31\u6307\u5b9a\u7684\u7b2c\u4e00\u500b\u8a3b\u91cb\u7684\u503c\u3002
-CWWKC1403.unsupported.tx.type.useraction=\u8acb\u66f4\u65b0\u61c9\u7528\u7a0b\u5f0f\u4ee5\u5207\u63db\u81f3\u5176\u4e2d\u4e00\u500b\u652f\u63f4\u7684\u503c\uff0c\u6216\u79fb\u9664\u5176\u4e2d\u4e00\u500b\u8a3b\u91cb\u3002
