+++ dev/io.openliberty.concurrent.internal.cdi.4.0/src/io/openliberty/concurrent/internal/cdi4/interceptor/AsyncInterceptor.java
@@ -24,12 +24,11 @@ import java.util.concurrent.RejectedExecutionException;
 import javax.naming.InitialContext;
 import javax.naming.NamingException;
 
-import com.ibm.websphere.ras.Tr;
-import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.websphere.ras.annotation.Trivial;
 import com.ibm.ws.concurrent.WSManagedExecutorService;
 import com.ibm.ws.ffdc.annotation.FFDCIgnore;
 
+import io.openliberty.concurrent.internal.messages.ConcurrencyNLS;
 import jakarta.annotation.Priority;
 import jakarta.enterprise.concurrent.Asynchronous;
 import jakarta.interceptor.AroundInvoke;
@@ -44,17 +43,15 @@ import jakarta.transaction.Transactional.TxType;
 public class AsyncInterceptor implements Serializable {
     private static final long serialVersionUID = 7447792334053278336L;
 
-    private static final TraceComponent tc = Tr.register(AsyncInterceptor.class);
-
     @AroundInvoke
     @FFDCIgnore({ ClassCastException.class, NamingException.class }) // application errors raised directly to the app
     public Object intercept(InvocationContext invocation) throws Exception {
         Method method = invocation.getMethod();
         validateTransactional(method);
         if (method.getDeclaringClass().getAnnotation(Asynchronous.class) != null)
-            throw new UnsupportedOperationException(Tr.formatMessage(tc, "CWWKC1401.class.anno.disallowed",
-                                                                     "@Asynchronous",
-                                                                     method.getDeclaringClass().getName()));
+            throw new UnsupportedOperationException(ConcurrencyNLS.getMessage("CWWKC1401.class.anno.disallowed",
+                                                                              "@Asynchronous",
+                                                                              method.getDeclaringClass().getName()));
 
         Asynchronous anno = method.getAnnotation(Asynchronous.class);
 
@@ -63,12 +60,12 @@ public class AsyncInterceptor implements Serializable {
         if (!returnType.equals(CompletableFuture.class)
             && !returnType.equals(CompletionStage.class)
             && !returnType.equals(Void.TYPE)) // void
-            throw new UnsupportedOperationException(Tr.formatMessage(tc, "CWWKC1400.unsupported.return.type",
-                                                                     returnType,
-                                                                     method.getName(),
-                                                                     method.getClass().getName(),
-                                                                     "@Asynchronous",
-                                                                     "[CompletableFuture, CompletionStage, void]"));
+            throw new UnsupportedOperationException(ConcurrencyNLS.getMessage("CWWKC1400.unsupported.return.type",
+                                                                              returnType,
+                                                                              method.getName(),
+                                                                              method.getClass().getName(),
+                                                                              "@Asynchronous",
+                                                                              "[CompletableFuture, CompletionStage, void]"));
 
         Object executor;
         try {
@@ -86,13 +83,13 @@ public class AsyncInterceptor implements Serializable {
                 for (Class<?> i : executor.getClass().getInterfaces())
                     interfaces.add(i.getName());
 
-            throw new RejectedExecutionException(Tr.formatMessage(tc, "CWWKC1402.not.managed.executor",
-                                                                  "@Asynchronous",
-                                                                  method.getName(),
-                                                                  method.getClass().getName(),
-                                                                  anno.executor(),
-                                                                  executor.getClass().getName(),
-                                                                  interfaces), x);
+            throw new RejectedExecutionException(ConcurrencyNLS.getMessage("CWWKC1402.not.managed.executor",
+                                                                           "@Asynchronous",
+                                                                           method.getName(),
+                                                                           method.getClass().getName(),
+                                                                           anno.executor(),
+                                                                           executor.getClass().getName(),
+                                                                           interfaces), x);
         }
 
         return managedExecutor.newAsyncMethod(this::invoke, invocation);
@@ -139,13 +136,13 @@ public class AsyncInterceptor implements Serializable {
                 case REQUIRES_NEW:
                     break;
                 default:
-                    throw new UnsupportedOperationException(Tr.formatMessage(tc, "CWWKC1403.unsupported.tx.type",
-                                                                             "@Transactional",
-                                                                             tx.value(),
-                                                                             "@Asynchronous",
-                                                                             method.getName(),
-                                                                             method.getDeclaringClass().getName(),
-                                                                             Arrays.asList(TxType.REQUIRES_NEW, TxType.NOT_SUPPORTED)));
+                    throw new UnsupportedOperationException(ConcurrencyNLS.getMessage("CWWKC1403.unsupported.tx.type",
+                                                                                      "@Transactional",
+                                                                                      tx.value(),
+                                                                                      "@Asynchronous",
+                                                                                      method.getName(),
+                                                                                      method.getDeclaringClass().getName(),
+                                                                                      Arrays.asList(TxType.REQUIRES_NEW, TxType.NOT_SUPPORTED)));
             }
     }
 }
