+++ dev/io.openliberty.concurrent.internal.cdi/bnd.bnd
@@ -13,10 +13,10 @@
 -include= ~../cnf/resources/bnd/bundle.props
 bVersion=1.0
 
-javac.source: 11
-javac.target: 11
+javac.source: 21
+javac.target: 21
 
-Require-Capability: osgi.ee; filter:="(&(osgi.ee=JavaSE)(version=11))"
+Require-Capability: osgi.ee; filter:="(&(osgi.ee=JavaSE)(version=21))"
 
 Bundle-Name: Interceptors for Jakarta Concurrency
 Bundle-SymbolicName: io.openliberty.concurrent.internal.cdi
@@ -24,11 +24,8 @@ Bundle-Description: Provides interceptors for Jakarta Concurrency; version=${bVe
 
 WS-TraceGroup: concurrent
 
-# TODO: these CDI package imports can be removed when JEE9 compatibility is no longer required
 Import-Package: \
- jakarta.enterprise.event;version="[3.0,5)",\
- jakarta.enterprise.inject.spi;version="[3.0,5)",\
- *
+  *
 
 Private-Package: \
   io.openliberty.concurrent.internal.cdi.*,\
@@ -52,6 +49,7 @@ instrument.classesExcludes: io/openliberty/concurrent/internal/cdi/resources/*.c
   com.ibm.ws.javaee.version,\
   com.ibm.ws.kernel.service,\
   com.ibm.ws.threading,\
+  io.openliberty.concurrent.internal,\
   io.openliberty.jakarta.annotation.2.1,\
   io.openliberty.jakarta.cdi.4.1,\
   io.openliberty.jakarta.concurrency.3.1,\
